GUI (#54)
* Healthbar Object added

The Player now has an Healthbar Object, but yet without function

* added the healthbar

the healthbar concept works, however It doesnt connect with an Object for some Reason.
Still needs to be fixed.

* Compass Bar/ Task List Added

Old concept of CompassBar removed, and a new Version is implemented

The UI for a Task List is added too.

* Documentation Added, old Files removed

the scripts are now explained and and old Compass Bar concept has been removed.

* Clean Up

unnecessary Debugs deleted, and HealthScript back to the "old" concept

* Remove ui elements from player prefab

* Format file

* Delete TextmeshPro examples

* Delete unnecessary meta file.

* Change comments

* Remove healthbar from health.cs

* Change comment style

Co-authored-by: socke300 <60921761+socke300@users.noreply.github.com>