Map assets (#64)
* Advanced Player Movement (#43)

* Add Sprinting

Add the ability to sprint (with stamina + cooldown)

* Add Crouching

Add the ability to crouch

* Sync player scale between clients

Enabled synchronization of players scale between clients

* Minor changes

- Add prone stance, not using it due to bug

- Enable scale interpolation via network

- Sprinting cancels crouch

* format file

Co-authored-by: socke300 <60921761+socke300@users.noreply.github.com>

* GUI (#44)

* Healthbar Object added

The Player now has an Healthbar Object, but yet without function

* added the healthbar

the healthbar concept works, however It doesnt connect with an Object for some Reason.
Still needs to be fixed.

* Compass Bar/ Task List Added

Old concept of CompassBar removed, and a new Version is implemented

The UI for a Task List is added too.

* Documentation Added, old Files removed

the scripts are now explained and and old Compass Bar concept has been removed.

* Clean Up

unnecessary Debugs deleted, and HealthScript back to the "old" concept

* Copy file because of merge conflict.

Co-authored-by: socke300 <60921761+socke300@users.noreply.github.com>

* placing assets and set atmosphere

* fixing merge errors2

Co-authored-by: yori21 <alex.ochs@mni.thm.de>
Co-authored-by: socke300 <60921761+socke300@users.noreply.github.com>
Co-authored-by: konstyy <64797633+konstyy@users.noreply.github.com>