GUI-Advanced (#52)
* Started Player UI

UI prefab and Crosshair picture was added

* Compass sprites added

The Sprites for the Compass are done and loaded into the Images Folder.
I also started a Compass Script.

* HealthbarFrame Updated

The Healthbarframe is updated.
Started a script for the slider

* Healthbar Script added

A Healthbar Script is added and implemented in the Health script of the player.

* Fixed the compass

Compass is fixed, Healthbar Sprite Updated

* Added Weapon GUI

Weapon GUI  is added, still needs to be connected to the Weapon Scripts

* Created Sprites for Weapons

Made Sprites out of the Prefabs of the 4 Weapons

* Sprites reworked

The healthbar and weapon HUD-sprites are reworked, and improved.

* HUD Scaling fixed

the HUD should now Scale with screensize

* Small fix HUD

TaskList and compass moved a bit.

* Comments added

Comments to all Scripts have been updated.

* Update PlayerHUD.cs

Minor Comment Adjustment

* Script Commenting updated

The comments in all UI related scripts have been updated to the new standard.

* Format file

* Change comments

Co-authored-by: socke300 <60921761+socke300@users.noreply.github.com>