GUI changes traced back (#116)
* GUI changes traced back

Minimap, Inventory and Healthbar are added on a new branch to attempt an error fix when merging.

* Update UI TEST.unity

* Fix Whitepsacing

* Update PlayerCanvas.cs

* Update PlayerCanvas.cs

* Update PlayerCanvas.cs

* Update PlayerCanvas.cs

* Documentation Update

* Update Minimap.cs

* Problems Fixed:

1. Multiplayer Overlap: Fixed
2. Healthbar:Fixed
3. Minimap: In Progress
4. Multiple Network Identities Fixed

The entire UI is now on the animated model.

* Update Health.cs

* Update GUIAnimated.prefab

The minimap camera is repositioned.
It now serves as a way to look behind you, but not showing the entire map.

* Minimap Fixed

The minimap now does not render anything tagged as:
UI
Monster
Player
MeleeWeapon
Celling
Device

To try it out use the "GUIAnimated" Prefab in any scene.

* Fixes whitespaces

Co-authored-by: Dominik <dominik.muench@mni.thm.de>